requirements.txt
asgiref==3.7.2
astroid==3.0.1
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
decouple==0.0.7
dill==0.3.7
distlib==0.3.8
Django==5.0.2
django-bootstrap5==23.4
django-ckeditor==6.7.1
django-filter==24.2
django-grappelli==3.0.8
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
djangorestframework==3.15.1
Eel==0.16.0
filelock==3.13.1
Flask==3.0.2
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.26.4
phonenumbers==8.13.35
phonenumberslite==8.13.35
Pillow==10.1.0
platformdirs==4.0.0
psycopg2==2.9.9
pycparser==2.21
pyfiglet==1.0.2
pylint==3.0.2
pyparsing==3.1.2
pypiwin32==223
pypng==0.20220715.0
python-decouple==3.8
pyttsx3==2.90
pytube==15.0.0
pywin32==306
qrcode==7.4.2
sqlparse==0.4.4
termcolor==2.3.0
tomlkit==0.12.3
ttkbootstrap==1.10.1
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.1
Werkzeug==3.0.1
whichcraft==0.6.1
windows-curses==2.3.2
zope.event==5.0
zope.interface==6.2
